+++ docs/README.md
@@ -8,7 +8,7 @@
 | [math/activations.md](math/activations.md) | Activations ↔ `src/networks/activation/` |
 | [math/learning_rates.md](math/learning_rates.md) | LR schedules ↔ `src/networks/learning_rate/` |
 | [math/metrics.md](math/metrics.md) | Training metrics ↔ `src/networks/metrics.py` |
-| [math/datasets.md](math/datasets.md) | Datasets ↔ `src/data/dataset.py` + CLI `-t`; run `config.json` `dataset` block for saved runs |
+| [math/datasets.md](math/datasets.md) | Datasets ↔ `src/data/dataset.py` hub + `*_dataset.py` modules + CLI `-t`; run `config.json` `dataset` block for saved runs |
 | [streamlit_app.md](streamlit_app.md) | Streamlit UI (`src/streamlit_main.py`, `src/streamlit_app/`) |
 
 **Agent-oriented detail:** `.cursor/` (architecture, CLI patterns, file layout).
